Add author and editor to notes
This commit adds the author and editor ID and name to the notes in the
database. This feature is needed for the UI component update.

We use the author_id as fallback for the user property when no user property is set,
this is to assure that the old UI component still works. The same is for old messages that only have
a user set, we use this as the author_id, so old message seem to the UI component as a new message.