Add remove compatibility with new service messages
Prior to this commit we only had backwards compatibility with the deleting
of old service messages. The new messages don't have the `user` property set,
but the `author_id`. In order to support both the new and old messages, we
changed the filter to a `or` statement where the author or the user id needs
to match the current context user.