Fix unable to delete message from other users
Prior to this commit we check if the user had permission to delete a message
and after that we also filter the message you want to delete on the user.

But because the filter was only for your own messages you could not delete
messages of other users.

As a user without MANAGE_AGENT permission can only delete it's own messages
we probably added that filter in the past for this use case, but after adding
that a user with MANAGE_AGENT could remove messages of other users we should have
deleted that filter. The check_permission method already checks if you may or
may not delete the message anyways.